Move `Player` class to separate file Fixes #53
