+++ bot.py
@@ -1,4 +1,6 @@
-import os, random, string, logging, jsonpickle
+import os, random, string, logging
+
+from player import Player
 
 from discord.ext import commands
 
@@ -45,80 +47,6 @@ async def on_ready():
     print('We have logged in as ')
     print(bot.user.display_name)
     print('\nLet''s make some words')
-
-
-class Player:
-
-    def __init__(self, user):
-        
-        self.name = format_name(user)
-        self.statefile = f".lws/{self.name}.json"
-        
-        try:
-            with open(self.statefile, 'r') as statefile:
-                self.state = jsonpickle.decode(statefile.read())
-        except FileNotFoundError:
-            logging.debug(f"statefile not found; initializing statefile for {self.name}")
-            self.state = {}
-            self.state["username"] = user.name
-            self.state["letters"] = []
-            self.state["score"] = 0
-            self.save_state()
-
-    def get_letters(self):
-        return self.state["letters"]
-
-    def add_letter(self, letter):
-        self.state["letters"].append(letter)
-        self.save_state()
-
-    def cheat(self):
-        try:
-            self.remove_all_letters()
-            for ltr in ["E", "A", "I", "S", "T", "L", "N", "R"]:
-                self.add_letter(ltr)
-            return("Your hand is now: E, A, I, S, T, L, N, and R!")
-        except:
-            logging.error("# Error 4 #: Error when cheating in letters")
-            return("Unable to help you cheat, cheaty!")
-
-    def remove_letter(self, letter):
-        self.state["letters"].remove(letter.upper())
-        self.save_state()
-
-    def remove_letters(self, letters):
-        for letter in letters:
-            self.remove_letter(letter)
-
-    def remove_all_letters(self):
-        self.state["letters"] = []
-        self.save_state()
-
-    def num_letters(self):
-        return len(self.state["letters"])
-
-    def get_username(self):
-        return self.state["username"]
-
-    def add_points(self, points):
-        self.state["score"] += points
-        self.save_state()
-
-    def get_score(self):
-        return self.state["score"]
-
-    def save_state(self):
-        pickled = jsonpickle.encode(self.state)
-        with open(self.statefile, 'w') as statefile:
-            statefile.write(pickled)
-            statefile.close()
-
-    def __str__(self):
-        return self.get_username()
-
-
-def format_name(user):
-    return f"{user.name}#{user.discriminator}"
 
 
 @bot.command(brief='Buy a new letter', description='For getting new letters')
@@ -172,7 +100,8 @@ async def word(ctx, *args):
                 logging.error(msg)
                 await ctx.send(msg)
     else:
-        logging.error("# Error 2 #: I don't know the word ""{}"" yet, sorry".format(word))
+        msg = f"# Error 2 #: I don't know the word '{word}' yet, sorry"
+        logging.info(msg)
 
 
 @bot.command(brief='Show me my progress', description='Get my score')
